% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_i_stat.R
\name{list_i_stat}
\alias{list_i_stat}
\title{Obtain the complete list of available data sets (source: I.Stat)}
\usage{
list_i_stat(lang = "ita")
}
\arguments{
\item{lang}{language parameter for labels ("ita" for Italian, "eng" for English)}
}
\value{
It returns the complete list of available data sets from I.Stat with their id and name.
}
\description{
Obtain the complete list of available data sets (source: I.Stat)
}
\examples{
\donttest{list_i_stat()}
}
