% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_istatdata.R
\name{list_istatdata}
\alias{list_istatdata}
\title{Obtain the complete list of available data sets (source: IstatData)}
\usage{
list_istatdata(lang = "ita")
}
\arguments{
\item{lang}{language parameter for labels ("ita" for Italian, "eng" for English)}
}
\value{
It returns the complete list of available data sets from IstatData with their agencyId, id, version and name.
}
\description{
Obtain the complete list of available data sets (source: IstatData)
}
\examples{
\donttest{list_istatdata()}
}
