% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interactive.R
\name{plot_interactive}
\alias{plot_interactive}
\title{Plot dataset interactively}
\usage{
plot_interactive(df)
}
\arguments{
\item{df}{data set as data.frame}
}
\value{
It returns the chosen plot.
}
\description{
Build different types of exploratory graphs (scatter plot, bar plot, pie chart). You can interactively choose the plot that you want create. Once you have chosen the plot, you can interactively choose the variables from the data set for which you want to build the plot.
}
\note{
plot_interactive allows you to have an idea about the general trend of your data, and it's intended to be used with exploratory purpose.
}
\examples{
if(interactive()) {
  plot_interactive(iris)
}
}
