% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyIstat.R
\name{shinyIstat}
\alias{shinyIstat}
\title{shinyIstat}
\usage{
shinyIstat()
}
\value{
It opens the app.
}
\description{
A graphic interface that makes searching, downloading and filtering data sets from Istat easier. Call shinyIstat() to get started. This shinyApp was built using the same functions of istat package but they have been adapted for the shiny. You will find additional information and help inside the app.
}
\note{
Calling a ShinyApp equals to calling a R function. For this reason, once called shinyIstat(), R will be busy processing it until the app will be closed. As a consequence, all other R functionalities can be processed only when the app is closed.
}
\examples{
if(interactive()) {
  shinyIstat()
}
}
